fixing zmp tests with namedtuple protocol
